requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pandas==2.1.3
numpy==1.25.2
reportlab==4.0.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
xlsxwriter==3.1.9
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1